+++ backend/src/app.js
@@ -1,5 +1,6 @@
 import express from "express";
 import healthRouter from "./routes/health.routes.js"
+import userRouter from "./routes/user.routes.js"
 
 
 const app = express();
@@ -7,7 +8,8 @@ const app = express();
 
 app.use(express.json());
 
-app.use("/api/v1/", healthRouter);
+app.use("/api/v1/health", healthRouter);
+app.use("/api/v1/users", userRouter);
 
 
 // middleware will go here