+++ backend/src/routes/health.routes.js
@@ -4,7 +4,7 @@ import express from "express";
 const router = express.Router();
 
 
-router.get("/health", (req, res)=>{
+router.get("/", (req, res)=>{
     res.status(200).json({
         "success": true,
         "message" : "server is running"